Enable linters in golansci-lint and fix issues
Enabled linters:
depguard
errchkjson
errorlint
exportloopref
goheader
goimports
importas
nolintlint
predeclared
stylecheck
unconvert
whitespace

+++ udp/client/clientobserve_test.go
@@ -93,7 +93,7 @@ func TestClientConnObserve(t *testing.T) {
 						defer cc.ReleaseMessage(req)
 						req.SetCode(codes.Content)
 						req.SetContentFormat(message.TextPlain)
-						req.SetObserve(uint32(i) + 2)
+						req.SetObserve(i + 2)
 						req.SetBody(p)
 						req.SetETag(etag)
 						req.SetToken(token)
@@ -303,7 +303,7 @@ func (o *observer) observe(req *pool.Message) {
 	}
 	obs, err := req.Observe()
 	require.NoError(o.t, err)
-	if obs > uint32(o.numEvents) {
+	if obs > o.numEvents {
 		select {
 		case o.done <- true:
 		default:
